Fix the output from this module
The ternary conditional evaluates every expression so the missing
resource causes it to fail, which happens silently in an `output`.

This uses the fact that the resources have a `count` to treat them
as lists, `concat` them to one list, `compact` them to remove empty
elements, then take the first (and only) element from the list.

JIRA: PLAT-1117